flowtop: Change flows layout to 1-row view
Changed flows list layout to look more a top-like output
with header and in 1 line. When -s option is specified
then layout changes to 2 lines view including with src peer
info and dst under it on next line.

Also shortified flow state names to allocate less space.

Removed presenter_get_port be cause ports are printed for both peers
separately.

The flow duration time is printed in very short form in one of the
units:
    XXd - days
    XXh - hours
    XXm - minutes
    XXs - seconds

the reason is that it is enough to have actually generic understanding
about flow time in the biggest time unit.

Signed-off-by: Vadim Kochan <vadim4j@gmail.com>