trafgen: proto: Allow to set field with variable length
It is quite tricky to set field value with a variable length
(i.e. DNS query name), to make it possible the field
needs to be added to header with 'len=0' in that case there
will be no any payload allocation, but only while setting the field
value the packet will be appended with a real length bytes and after
the field needs to be relocated to the right place.

Signed-off-by: Vadim Kochan <vadim4j@gmail.com>