flowtop: Improve and unify up/down scrolling
Move scrolling logic to the ui.c module, it requires to have
some data iteration provided in flowtop.c and delegated to ui.c part.

So approach is that now flowtop provides 2 additional callbacks for:

    1) Iterate over flows/procs list
    2) Draw flow/proc on each iteration which is controlled from ui.c

it allows to unify scrolling logic and delegate it to the ui.c, in the
future it should allow to easy handle press event on selected row and
drow some additional information, or draw a cursor line per selected
row.

Also fixed case when down scrolling was bigger that printed rows, not
it is handled by ui part.

Signed-off-by: Vadim Kochan <vadim4j@gmail.com>